Install R Markdown tutorial packages with Conda
Install all R Markdown tutorial software packages with Conda as the
default, rather than `install.packages()` or `BiocManager`. The thinking
is that the course teaches the students how to use Conda, so we should
try to use Conda in as many places as possible throughout the rest of
the course (post-Conda tutorial, that is).