Update tutorials/reproducibility overview figures
Remove the old tutorials overview figure, which didn't contribute much
knowledge to the reader, and replace it with the reproducibility
overview image with tool logos. This connects the various different
parts of reproducibility (data, code, workflow, *etc.*) with the tools
themselves, putting them in context and making it clearer to the reader.