Change Singularity instructions to run from Docker
Change the Singularity tutorial material to only run
singularity-in-docker, better reflecting the day-to-day work that the
teachers to regarding container-related work. Ideally it should be clear
that Singulary/Apptainer is still a great piece of software, especially
if working on HPC systems, but that Docker is preferable for
cross-platform compatibility and ease-of-use for most users.