Streamline `[slim/conda].Dockerfile` files
The instructions for starting to build Docker images were more
complicated than they needed to be for teaching the students the basics
of building images, and this commit streamlines them. Things that have
been removed include:

 - Remove redundant SHELL directive
 - Time zone packages
 - Port forwarding
 - Initialising Conda (not needed with Miniforge base)
 - Usage of `--set subdir linux-64` and `--platform linux/amd64`
   instructions, as they are not needed at this stage
 - Creating a new Conda environment and activating it instead of using
   the base environment

Hopefully this streamlining will make the very first entry into
Dockerfiles and their use easier to understand for the students.