Remove `tutorials/git` to avoid duplication
The `tutorials/git` directory contains copies and some of the files from
the container tutorial, but nothing else. The only thing the students
are doing with these files are copying them to another local Git
repository they are working on.

Keeping the copies in the Git directory up-to-date with the actual files
in the container directory is an extra step which we often forget. This
commit simply removes the Git tutorial directory and replaces the
relevant text in the Git page that they should copy the files from the
container tutorial instead.